changed structure yet again to suit convention, refactored and renamed some files too. Hopefully no more restructuring rounds.
