+++ config/config.js
@@ -1,5 +1,6 @@
 const MINE_RATE = 1000
 const INITIAL_DIFFICULTY = 3
+const STARTING_BALANCE = 100
 
 const GENESIS_DATA = {
   timestamp: 1,
@@ -13,4 +14,5 @@ const GENESIS_DATA = {
 module.exports = {
   GENESIS_DATA,
   MINE_RATE,
+  STARTING_BALANCE,
 }
