+++ package.json
@@ -11,12 +11,16 @@
     "start-redis": "redis-server --daemonize yes",
     "stop-redis": "redis-cli shutdown"
   },
+  "jest": {
+    "testEnvironment": "node"
+  },
   "keywords": [],
   "author": "Ed Leonard",
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.18.3",
     "cross-env": "^5.2.0",
+    "elliptic": "^6.4.1",
     "express": "^4.16.3",
     "hex-to-binary": "^1.0.1",
     "prettier": "^1.16.4",
