+++ src/util/crypto-hash.js
