+++ src/util/crypto-hash.unit.test.js
@@ -1,4 +1,4 @@
-const cryptoHash = require('../../src/helpers/crypto-hash')
+const cryptoHash = require('./crypto-hash')
 
 describe('cryptoHash()', () => {
   it('geneates a SHA-256 hashed output', () => {
