+++ src/util/elliptic.js
@@ -1,0 +1,7 @@
+const EC = require('elliptic').ec
+
+const ec = new EC('secp256k1')
+
+module.exports = {
+  ec,
+}
