+++ src/wallet/wallet.unit.test.js
@@ -1,0 +1,17 @@
+const Wallet = require('./wallet')
+
+describe('Wallet', () => {
+  let wallet
+
+  beforeEach(() => {
+    wallet = new Wallet()
+  })
+
+  it('has a `balance`', () => {
+    expect(wallet).toHaveProperty('balance')
+  })
+
+  it('has a `publicKey`', () => {
+    expect(wallet).toHaveProperty('publicKey')
+  })
+})
